+++ src/main/kotlin/com/exactpro/th2/processor/cache/collector/Factory.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.exactpro.th2.processor.cache.collecotor
+package com.exactpro.th2.processor.cache.collector
 
 import com.exactpro.th2.common.event.Event
 import com.exactpro.th2.processor.api.IProcessor
